// src/utils/models/Business.ts
export class GenericResponse{
    answer: string;
    answerList: string[];
    code: string;
    genericObject: any;
    variable: string;
}

export class GenericResponseList{
    answer: any;
    answerList: [];
    code: any;
    genericObject: [];
    variable: any;
}

export class notification{
    title: string;    
    body: string;    
}

export class to{
    to: string;   
}